Moved fires widgets to folder
