+++ app/javascript/components/widgets/fires/fires-alerts-cumulative/index.js
