+++ app/javascript/components/widgets/fires/fires-alerts-cumulative/selectors.js
