+++ app/javascript/components/widgets/fires/fires-alerts/index.js
