+++ app/javascript/components/widgets/fires/fires-alerts/selectors.js
