+++ app/javascript/components/widgets/fires/fires-ranked/index.js
