+++ app/javascript/components/widgets/fires/fires-within/index.js
