+++ app/javascript/components/widgets/fires/fires-within/selectors.js
