+++ app/javascript/components/widgets/fires/fires/index.js
