+++ app/javascript/components/widgets/fires/fires/selectors.js
