+++ app/javascript/components/widgets/fires/glad-alerts/index.js
