+++ app/javascript/components/widgets/fires/glad-ranked/index.js
