+++ app/javascript/components/widgets/fires/glad-ranked/selectors.js
