+++ app/javascript/components/widgets/fires/glads/index.js
