+++ app/javascript/components/widgets/fires/glads/selectors.js
